requirements.txt
amqp==5.3.1
asgiref==3.6.0
attrs==25.1.0
billiard==4.2.1
celery==5.4.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.32.0
colorama==0.4.6
dj-database-url==2.3.0
Django==4.2
django-cors-headers==3.14.0
django-environ==0.12.0
django-filter==23.2
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.28.0
drf-yasg==1.21.8
gunicorn==23.0.0
h11==0.14.0
idna==3.4
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
packaging==24.2
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
PyJWT==2.6.0
python-dateutil==2.9.0.post0
pytz==2023.3
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.30.0
rpds-py==0.22.3
setuptools==75.8.0
six==1.16.0
sqlparse==0.4.4
stripe==5.4.0
typing_extensions==4.12.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
